+++ flandmark/cppbuild.sh
@@ -1,30 +1,66 @@
-#!/bin/sh
 if [[ -z "$PLATFORM" ]]; then
     echo "This file is meant to be included by the parent cppbuild.sh script"
     exit 1
 fi
 
-# FLANDMARK_VERSION=a0981a3b09cc5534255dc1dcdae2179097231bdd
-# download https://github.com/uricamic/flandmark/archive/$FLANDMARK_VERSION.zip flandmark.zip
-
-# unzip -o flandmark.zip -d flandmark/sources
+FLANDMARK_VERSION=master
+download https://github.com/uricamic/flandmark/archive/$FLANDMARK_VERSION.zip flandmark-$FLANDMARK_VERSION.zip
 
 mkdir -p $PLATFORM
 cd $PLATFORM
+INSTALL_PATH=`pwd`
+mkdir -p include lib bin
+unzip -o ../flandmark-$FLANDMARK_VERSION.zip
+cd flandmark-$FLANDMARK_VERSION
 
-mkdir -p include
-cp -r ../flandmark/sources/libflandmark/flandmark_detector.h include
-cp -r ../flandmark/sources/libflandmark/msvc-compat.h include
-cp -r ../../../opencv/cppbuild/$PLATFORM/include/opencv/cv.h include
-cp -r ../../../opencv/cppbuild/$PLATFORM/include/opencv/cvaux.h include
-cp -r ../../../opencv/cppbuild/$PLATFORM/include/opencv2/ include
+OPENCV_PATH=$INSTALL_PATH/../../../opencv/cppbuild/$PLATFORM/
 
-mkdir -p bin
-cp -r ../flandmark/build/x64/libflandmark/RelWithDebInfo/flandmark_shared.dll bin
-cp -r ../../../opencv/cppbuild/$PLATFORM/bin/* bin
+case $PLATFORM in
+    android-arm)
+        cmake -DCMAKE_TOOLCHAIN_FILE=$INSTALL_PATH/../../android-arm.cmake -DCMAKE_BUILD_TYPE=Release -DOpenCV_DIR=$OPENCV_PATH/sdk/native/jni/ -DANDROID_NDK_ABI_NAME=armeabi_v7a
+        make -j4 flandmark_static
+        cp libflandmark/*.h ../include
+        cp libflandmark/*.a ../lib
+        ;;
+    android-x86)
+        cmake -DCMAKE_TOOLCHAIN_FILE=$INSTALL_PATH/../../android-x86.cmake -DCMAKE_BUILD_TYPE=Release -DOpenCV_DIR=$OPENCV_PATH/sdk/native/jni/ -DANDROID_NDK_ABI_NAME=x86
+        make -j4 flandmark_static
+        cp libflandmark/*.h ../include
+        cp libflandmark/*.a ../lib
+        ;;
+    linux-x86)
+        CC="gcc -m32" CXX="g++ -m32" cmake -DCMAKE_BUILD_TYPE=Release -DOpenCV_DIR=$OPENCV_PATH/share/OpenCV/
+        make -j4 flandmark_static
+        cp libflandmark/*.h ../include
+        cp libflandmark/*.a ../lib
+        ;;
+    linux-x86_64)
+        CC="gcc -m64" CXX="g++ -m64" cmake -DCMAKE_BUILD_TYPE=Release -DOpenCV_DIR=$OPENCV_PATH/share/OpenCV/
+        make -j4 flandmark_static
+        cp libflandmark/*.h ../include
+        cp libflandmark/*.a ../lib
+        ;;
+    macosx-x86_64)
+        cmake -DCMAKE_BUILD_TYPE=Release -DOpenCV_DIR=$OPENCV_PATH/share/OpenCV/
+        make -j4 flandmark_static
+        cp libflandmark/*.h ../include
+        cp libflandmark/*.a ../lib
+        ;;
+    windows-x86)
+        cmake -DCMAKE_BUILD_TYPE=Release -DOpenCV_DIR=$OPENCV_PATH/opencv/build/
+        nmake flandmark_static
+        cp libflandmark/*.h ../include
+        cp libflandmark/*.lib ../lib
+        ;;
+    windows-x86_64)
+        cmake -DCMAKE_BUILD_TYPE=Release -DOpenCV_DIR=$OPENCV_PATH/opencv/build/
+        nmake flandmark_static
+        cp libflandmark/*.h ../include
+        cp libflandmark/*.lib ../lib
+        ;;
+    *)
+        echo "Error: Platform \"$PLATFORM\" is not supported"
+        ;;
+esac
 
-mkdir -p lib
-cp -r ../flandmark/build/x64/libflandmark/RelWithDebInfo/flandmark_static.lib lib
-cp -r ../../../opencv/cppbuild/$PLATFORM/lib/* lib
-
-cd ..
+cd ../..