+++ flandmark/src/main/java/org/bytedeco/javacpp/presets/flandmark.java
@@ -26,29 +26,12 @@ import org.bytedeco.javacpp.tools.Info;
 import org.bytedeco.javacpp.tools.InfoMap;
 import org.bytedeco.javacpp.tools.InfoMapper;
 
-
 /**
  * @author Jarek Sacha
  */
-@Properties(target = "org.bytedeco.javacpp.flandmark", value = {
-        @Platform(value = "windows", preload = {"msvcr100", "msvcp100"},
-                link = {"opencv_core249", "opencv_imgproc249", "flandmark_static"},
-                include = {"<flandmark_detector.h>"}),
-        @Platform(value = "windows-x86",
-                includepath = {"../../opencv/cppbuild/windows-x86/include/opencv/",
-                        "../../opencv/cppbuild/windows-x86/include/opencv2/"},
-                linkpath = {"../../opencv/cppbuild/windows-x86/lib/"},
-                preloadpath = {"../../opencv/cppbuild/windows-x86/bin/",
-                        "C:/Program Files (x86)/Microsoft Visual Studio 10.0/VC/redist/x86/Microsoft.VC100.CRT/"}),
-        @Platform(value = "windows-x86_64",
-                includepath = {"../../opencv/cppbuild/windows-x86_64/include/opencv/",
-                        "../../opencv/cppbuild/windows-x86_64/include/opencv2/"},
-                linkpath = {"../../opencv/cppbuild/windows-x86_64/lib/"},
-                preloadpath = {"../../opencv/cppbuild/windows-x86_64/bin/",
-                        "C:/Program Files (x86)/Microsoft Visual Studio 10.0/VC/redist/x64/Microsoft.VC100.CRT/"})})
+@Properties(value = @Platform(include = {"flandmark_detector.h", "liblbp.h"}, link = "flandmark_static"),
+    inherit = opencv_imgproc.class, target = "org.bytedeco.javacpp.flandmark")
 public class flandmark implements InfoMapper {
     public void map(InfoMap infoMap) {
-        infoMap.put(new Info("flandmark_detector.cpp").skip())
-                .put(new Info().javaText("import static org.bytedeco.javacpp.opencv_core.*;"));
     }
 }